move otel instrumentation to logging dir
