+++ app/domain/shipper/metrics.go
@@ -4,7 +4,7 @@
 package shipper
 
 import (
-	"github.com/cloudzero/cloudzero-agent/app/instr"
+	"github.com/cloudzero/cloudzero-agent/app/logging/instr"
 	"github.com/prometheus/client_golang/prometheus"
 )
