+++ app/domain/shipper/shipper.go
@@ -17,8 +17,8 @@ import (
 	"time"
 
 	config "github.com/cloudzero/cloudzero-agent/app/config/gator"
-	"github.com/cloudzero/cloudzero-agent/app/instr"
 	"github.com/cloudzero/cloudzero-agent/app/lock"
+	"github.com/cloudzero/cloudzero-agent/app/logging/instr"
 	"github.com/cloudzero/cloudzero-agent/app/parallel"
 	"github.com/cloudzero/cloudzero-agent/app/storage/disk"
 	"github.com/cloudzero/cloudzero-agent/app/types"
