+++ app/logging/instr/span_test.go
